+++ scrapyd/runner.py
@@ -1,6 +1,5 @@
 import os
 import shutil
-import sys
 import tempfile
 from contextlib import contextmanager