+++ scrapyd/sqlite.py
@@ -16,12 +16,12 @@ class JsonSqliteDict(MutableMapping):
         self.table = table
         # about check_same_thread: http://twistedmatrix.com/trac/ticket/4040
         self.conn = sqlite3.connect(self.database, check_same_thread=False)
-        sql = "CREATE TABLE IF NOT EXISTS %s (key blob PRIMARY KEY, value blob)" % table
+        sql = f"CREATE TABLE IF NOT EXISTS {table} (key blob PRIMARY KEY, value blob)"
         self.conn.execute(sql)
 
     def __getitem__(self, key):
         key = self.encode(key)
-        sql = "SELECT value FROM %s WHERE key = ?" % self.table
+        sql = f"SELECT value FROM {self.table} WHERE key = ?"
         value = self.conn.execute(sql, (key,)).fetchone()
         if value:
             return self.decode(value[0])
@@ -29,39 +29,39 @@ class JsonSqliteDict(MutableMapping):
 
     def __setitem__(self, key, value):
         key, value = self.encode(key), self.encode(value)
-        sql = "INSERT OR REPLACE INTO %s (key, value) VALUES (?, ?)" % self.table
+        sql = f"INSERT OR REPLACE INTO {self.table} (key, value) VALUES (?, ?)"
         self.conn.execute(sql, (key, value))
         self.conn.commit()
 
     def __delitem__(self, key):
         key = self.encode(key)
-        sql = "DELETE FROM %s WHERE key = ?" % self.table
+        sql = f"DELETE FROM {self.table} WHERE key = ?"
         self.conn.execute(sql, (key,))
         self.conn.commit()
 
     def __len__(self):
-        sql = "SELECT COUNT(*) FROM %s" % self.table
+        sql = f"SELECT COUNT(*) FROM {self.table}"
         return self.conn.execute(sql).fetchone()[0]
 
     def __iter__(self):
         yield from self.iterkeys()
 
     def iterkeys(self):
-        sql = "SELECT key FROM %s" % self.table
+        sql = f"SELECT key FROM {self.table}"
         return (self.decode(row[0]) for row in self.conn.execute(sql))
 
     def keys(self):
         return list(self.iterkeys())
 
     def itervalues(self):
-        sql = "SELECT value FROM %s" % self.table
+        sql = f"SELECT value FROM {self.table}"
         return (self.decode(row[0]) for row in self.conn.execute(sql))
 
     def values(self):
         return list(self.itervalues())
 
     def iteritems(self):
-        sql = "SELECT key, value FROM %s" % self.table
+        sql = f"SELECT key, value FROM {self.table}"
         return ((self.decode(row[0]), self.decode(row[1])) for row in self.conn.execute(sql))
 
     def items(self):
@@ -84,23 +84,24 @@ class JsonSqlitePriorityQueue:
         self.table = table
         # about check_same_thread: http://twistedmatrix.com/trac/ticket/4040
         self.conn = sqlite3.connect(self.database, check_same_thread=False)
-        sql = "CREATE TABLE IF NOT EXISTS %s (id integer PRIMARY KEY, priority real key, message blob)" % table
+        sql = f"CREATE TABLE IF NOT EXISTS {table} (id integer PRIMARY KEY, priority real key, message blob)"
         self.conn.execute(sql)
 
     def put(self, message, priority=0.0):
         args = (priority, self.encode(message))
-        sql = "INSERT INTO %s (priority, message) VALUES (?, ?)" % self.table
+        sql = f"INSERT INTO {self.table} (priority, message) VALUES (?, ?)"
         self.conn.execute(sql, args)
         self.conn.commit()
 
     def pop(self):
-        sql = "SELECT id, message FROM %s ORDER BY priority DESC LIMIT 1" % self.table
-        idmsg = self.conn.execute(sql).fetchone()
-        if idmsg is None:
-            return
-        id, msg = idmsg
-        sql = "DELETE FROM %s WHERE id = ?" % self.table
-        c = self.conn.execute(sql, (id,))
+        sql = f"SELECT id, message FROM {self.table} ORDER BY priority DESC LIMIT 1"
+        id_message = self.conn.execute(sql).fetchone()
+        if id_message is None:
+            return id_message
+
+        _id, msg = id_message
+        sql = f"DELETE FROM {self.table} WHERE id = ?"
+        c = self.conn.execute(sql, (_id,))
         if not c.rowcount:  # record vanished, so let's try again
             self.conn.rollback()
             return self.pop()
@@ -108,12 +109,12 @@ class JsonSqlitePriorityQueue:
         return self.decode(msg)
 
     def remove(self, func):
-        sql = "SELECT id, message FROM %s" % self.table
+        sql = f"SELECT id, message FROM {self.table}"
         n = 0
-        for id, msg in self.conn.execute(sql):
+        for _id, msg in self.conn.execute(sql):
             if func(self.decode(msg)):
-                sql = "DELETE FROM %s WHERE id = ?" % self.table
-                c = self.conn.execute(sql, (id,))
+                sql = f"DELETE FROM {self.table} WHERE id = ?"
+                c = self.conn.execute(sql, (_id,))
                 if not c.rowcount:  # record vanished, so let's try again
                     self.conn.rollback()
                     return self.remove(func)
@@ -122,15 +123,15 @@ class JsonSqlitePriorityQueue:
         return n
 
     def clear(self):
-        self.conn.execute("DELETE FROM %s" % self.table)
+        self.conn.execute(f"DELETE FROM {self.table}")
         self.conn.commit()
 
     def __len__(self):
-        sql = "SELECT COUNT(*) FROM %s" % self.table
+        sql = f"SELECT COUNT(*) FROM {self.table}"
         return self.conn.execute(sql).fetchone()[0]
 
     def __iter__(self):
-        sql = "SELECT message, priority FROM %s ORDER BY priority DESC" % self.table
+        sql = f"SELECT message, priority FROM {self.table} ORDER BY priority DESC"
         return ((self.decode(message), priority) for message, priority in self.conn.execute(sql))
 
     def encode(self, obj):
@@ -149,14 +150,14 @@ class SqliteFinishedJobs:
         # about check_same_thread: http://twistedmatrix.com/trac/ticket/4040
         self.conn = sqlite3.connect(self.database, check_same_thread=False)
         sql = (
-            "CREATE TABLE IF NOT EXISTS %s (id integer PRIMARY KEY, "
-            "project text, spider text, job text, start_time datetime, end_time datetime)" % table
+            f"CREATE TABLE IF NOT EXISTS {table} "
+            "(id integer PRIMARY KEY, project text, spider text, job text, start_time datetime, end_time datetime)"
         )
         self.conn.execute(sql)
 
     def add(self, job):
         args = (job.project, job.spider, job.job, job.start_time, job.end_time)
-        sql = "INSERT INTO %s (project, spider, job, start_time, end_time) VALUES (?, ?, ?, ?, ?)" % self.table
+        sql = f"INSERT INTO {self.table} (project, spider, job, start_time, end_time) VALUES (?, ?, ?, ?, ?)"
         self.conn.execute(sql, args)
         self.conn.commit()
 
@@ -166,20 +167,17 @@ class SqliteFinishedJobs:
             limit = len(self) - finished_to_keep
             if limit <= 0:
                 return  # nothing to delete
-            w = "WHERE id <= " "(SELECT max(id) FROM (SELECT id FROM %s ORDER BY end_time LIMIT %d))" % (
-                self.table,
-                limit,
-            )
-        sql = "DELETE FROM {} {}".format(self.table, w)
+            w = f"WHERE id <= (SELECT max(id) FROM (SELECT id FROM {self.table} ORDER BY end_time LIMIT {limit}))"
+        sql = f"DELETE FROM {self.table} {w}"
         self.conn.execute(sql)
         self.conn.commit()
 
     def __len__(self):
-        sql = "SELECT COUNT(*) FROM %s" % self.table
+        sql = f"SELECT COUNT(*) FROM {self.table}"
         return self.conn.execute(sql).fetchone()[0]
 
     def __iter__(self):
-        sql = "SELECT project, spider, job, start_time, end_time FROM %s ORDER BY end_time DESC" % self.table
+        sql = f"SELECT project, spider, job, start_time, end_time FROM {self.table} ORDER BY end_time DESC"
         return (
             (
                 j[0],